requirements.txt
# Основные зависимости
numpy==1.26.4
pandas==2.2.1
tensorflow==2.15.0
matplotlib==3.8.3
scikit-learn==1.4.0

# Биоинформатические библиотеки
biopython==1.83
pyvcf==0.6.8  # Если будете работать с VCF-файлами

# Дополнительные утилиты
tqdm==4.66.1  # Для прогресс-баров (опционально)